Start moving code out of the basic functions file
