+++ include/qmckl_ao.h
@@ -13,15 +13,15 @@
 #include "qmckl_context.h"
 #include "qmckl_memory.h"
 
-qmckl_exit_code qmckl_compute_ao_basis_shell_gaussian_vgl_device(
+qmckl_exit_code_device qmckl_compute_ao_basis_shell_gaussian_vgl_device(
 	qmckl_context_device context, int prim_num, int shell_num, int point_num,
 	int nucl_num, int64_t *nucleus_shell_num, int64_t *nucleus_index,
 	double *nucleus_range, int64_t *shell_prim_index, int64_t *shell_prim_num,
 	double *coord, double *nucl_coord, double *expo, double *coef_normalized,
 	double *shell_vgl);
 
-qmckl_exit_code qmckl_compute_ao_vgl_gaussian_device(
-	const qmckl_context context, const int64_t ao_num, const int64_t shell_num,
+qmckl_exit_code_device qmckl_compute_ao_vgl_gaussian_device(
+	const qmckl_context_device context, const int64_t ao_num, const int64_t shell_num,
 	const int64_t point_num, const int64_t nucl_num,
 	const double *restrict coord, const double *restrict nucl_coord,
 	const int64_t *restrict nucleus_index,
@@ -30,8 +30,8 @@ qmckl_exit_code qmckl_compute_ao_vgl_gaussian_device(
 	const int32_t *restrict shell_ang_mom, const double *restrict ao_factor,
 	double *shell_vgl, double *restrict const ao_vgl);
 
-qmckl_exit_code qmckl_compute_ao_value_device(
-	const qmckl_context context, const int64_t ao_num, const int64_t shell_num,
+qmckl_exit_code_device qmckl_compute_ao_value_device(
+	const qmckl_context_device context, const int64_t ao_num, const int64_t shell_num,
 	const int64_t point_num, const int64_t nucl_num,
 	const double *restrict coord, const double *restrict nucl_coord,
 	const int64_t *restrict nucleus_index,
@@ -40,20 +40,20 @@ qmckl_exit_code qmckl_compute_ao_value_device(
 	const int32_t *restrict shell_ang_mom, const double *restrict ao_factor,
 	double *shell_vgl, double *restrict const ao_value);
 
-qmckl_exit_code
+qmckl_exit_code_device
 qmckl_provide_ao_basis_ao_vgl_device(qmckl_context_device context);
 
-qmckl_exit_code
+qmckl_exit_code_device
 qmckl_provide_ao_basis_shell_vgl_device(qmckl_context_device context);
 
-qmckl_exit_code
+qmckl_exit_code_device
 qmckl_provide_ao_basis_ao_value_device(qmckl_context_device context);
 
-qmckl_exit_code qmckl_get_ao_basis_ao_vgl_device(qmckl_context_device context,
+qmckl_exit_code_device qmckl_get_ao_basis_ao_vgl_device(qmckl_context_device context,
 												 double *const ao_vgl,
 												 const int64_t size_max);
 
-qmckl_exit_code qmckl_get_ao_basis_ao_value_device(qmckl_context_device context,
+qmckl_exit_code_device qmckl_get_ao_basis_ao_value_device(qmckl_context_device context,
 												   double *const ao_vgl,
 												   const int64_t size_max);
