+++ include/qmckl_mo.h
@@ -12,28 +12,28 @@
 #include "qmckl_context.h"
 #include "qmckl_memory.h"
 
-qmckl_exit_code qmckl_get_mo_basis_mo_vgl_device(qmckl_context context,
+qmckl_exit_code_device qmckl_get_mo_basis_mo_vgl_device(qmckl_context_device context,
 												 double *const mo_vgl,
 												 const int64_t size_max);
 
-qmckl_exit_code qmckl_get_mo_basis_mo_value_device(qmckl_context_device context,
+qmckl_exit_code_device qmckl_get_mo_basis_mo_value_device(qmckl_context_device context,
 												   double *mo_value,
 												   int64_t size_max);
 
-qmckl_exit_code
+qmckl_exit_code_device
 qmckl_get_mo_basis_mo_value_inplace_device(qmckl_context_device context,
 										   double *mo_value, int64_t size_max);
 
-qmckl_exit_code qmckl_provide_mo_basis_mo_value_device(qmckl_context context);
+qmckl_exit_code_device qmckl_provide_mo_basis_mo_value_device(qmckl_context_device context);
 
-qmckl_exit_code qmckl_provide_mo_basis_mo_vgl_device(qmckl_context context);
+qmckl_exit_code_device qmckl_provide_mo_basis_mo_vgl_device(qmckl_context_device context);
 
-qmckl_exit_code qmckl_compute_mo_basis_mo_value_device(
-	qmckl_context context, int64_t ao_num, int64_t mo_num, int64_t point_num,
+qmckl_exit_code_device qmckl_compute_mo_basis_mo_value_device(
+	qmckl_context_device context, int64_t ao_num, int64_t mo_num, int64_t point_num,
 	double *restrict coefficient_t, double *restrict ao_value,
 	double *restrict mo_value);
 
-qmckl_exit_code qmckl_compute_mo_basis_mo_vgl_device(
-	qmckl_context context, int64_t ao_num, int64_t mo_num, int64_t point_num,
+qmckl_exit_code_device qmckl_compute_mo_basis_mo_vgl_device(
+	qmckl_context_device context, int64_t ao_num, int64_t mo_num, int64_t point_num,
 	double *restrict coefficient_t, double *restrict ao_vgl,
 	double *restrict mo_vgl);
