+++ include/qmckl_trexio.h
@@ -12,6 +12,7 @@
 #include <string.h>
 #include <trexio.h>
 
+#include "qmckl_types.h"
 #include "qmckl_basic_functions.h"
 #include "qmckl_context.h"
 #include "qmckl_memory.h"
